+++ src/components/components/Breadcrumb/index.ts
@@ -1,6 +1,2 @@
 export { Breadcrumb } from './Breadcrumb'
-export type {
-  BreadcrumbProps,
-  BreadcrumbItem,
-  BreadcrumbLinkComponent,
-} from './Breadcrumb'
+export type { BreadcrumbProps, BreadcrumbItem } from './Breadcrumb'